+++ client/src/App.js
@@ -125,10 +125,10 @@ function App() {
                             <Route exact path = '/homework' component={Commons} />
                         }
                         {process.env.REACT_APP_ORG_ID === 'libretexts' &&
-                            <Route exact path = '/directory' component={Commons} />
+                            <Route exact path = '/libraries' component={Commons} />
                         }
                         {process.env.REACT_APP_ORG_ID === 'libretexts' &&
-                            <Route exact path = '/directory/:lib' component={Commons} />
+                            <Route exact path = '/libraries/:lib' component={Commons} />
                         }
                         {/* Standalone */}
                         {process.env.REACT_APP_ORG_ID === 'libretexts' &&